Containerized Compilation
The compilation process now occurs within a container that is equipped
with the appropriate GCC cross-compiler (specifically, GCC 13.2.0 and
binutils 2.41) targeted for i386-elf.

It ensures the compilation process is entirely portable and can be
executed on all Un*x systems or architecture (although further testing
is required). Moreover, it reduce the need of some flags and eliminates
the need of a dirty hack to compile on modern compiler. And it should
not break with future version of GCC or binutils.